moves AlertMessage up into its own directory
