+++ src/Flash/index.js
@@ -1,5 +1,5 @@
 import Flash from './Flash';
-import AlertMessage, { MessageTypes } from './AlertMessage';
+import AlertMessage, { MessageTypes } from '../Alert/AlertMessage';
 import useFlash from './useFlash';
 import withFlash, { withFlashPropTypes } from './withFlash';
